Split core and dialect-1 functionality of the hardware file parser.
